requirements.txt
openai
openai-whisper
ffmpeg